Adding scenes
There can be one active scene selected with core::set_scene(string);
All scenes have to be added in src/scenes.cpp

A scene is a group of the following functions in the same namespace:
    void init(int, char**, token)
    void deinit(token)
    void tick(u32, token)
    void render(token)
and optionally event handlers:
    void keydown(SDL_KeyboardEvent&, token)
    void keyup(SDL_KeyboardEvent&, token)
    void mouse_motion(SDL_MouseMotionEvent&, token)
    void mouse_wheel(SDL_MouseWheelEvent&, token)
    void mouseup(SDL_MouseButtonEvent&, token)
    void mousedown(SDL_MouseButtonEvent&, token)

The token is an unique pointer data type used only for that scene.
An example is a pointer to named struct within the scene's namespace.