utils: Adding trandom.hpp; Caching random distributions
Turns out i was using the <random> std library wrong.
I actually have to store my distributions for them to  have an actual effect.

Functions in random.hpp will cache the most used distributions.
random.hpp is deprecated and will be replaced with trandom.hpp
eventually.

Templates in trandom.hpp allow for easy creation of independent RNG
systems, with seeding abilities. Drawback is randomf not accepting
ranges but always using [0, 1) which should not be an issue.