game: Adjusting tile fall speed
Adding the concept of tiles falling slower when more tiles are falling
at once. The purpose is to make it easier to track what's happening on
the board.
Originally when i was splitting my focus to track multiple falling tiles
i was percepting them as falling faster while the speed was technically
the same. Adding this makes tiles fall a bit slower when a lot of them
are falling at once and if configured properly it can remove the feel of
different falling-speed that i was getting.

On side note, i had to consider if i want to base this slow-down on
count of falling tiles or lines. In conclusion tiles are the proper
choice, as the animation of falling actually splits them and there's
higher chance that the player will percept every one of them
individually.

The idea of using lines was that the filling of N blocks should be
taking constant time regardless vertical position of the blocks.
While technically true, a very slight difference here shouldn't impact
gameplay and the alternative feels better to me. I left it as a flag.

When the slower_massive_falls is enabled, the falling speed of tiles
gets multiplied by the slowness_per_falling_tile and the count of tiles
or lines. The slowness_per_falling_tile should be kept low.
It's a multiplier to keep it working independently of fall_duration.