utils: Adding font_cache.hpp
The font cache utilities reside inside the utils namespace.
It is a collection of functions therefore it stays in a namespace.

This system doesn't really provide all of the functionality the SDL2_ttf
library represents. It's purpose is to let you store a few fonts with
pre-set sizes. SDL2_ttf does support resizing, but caching the font a few
times should be faster in most common scenarios.

One can set a fixed limit of the number of stored fonts in FontCache
object, which is useful if working with multiple fonts.

The format of the cache entry makes comparison possible so it can be
enchanced to insert entries in order and use binary search when finding
cached fonts.