game: Activating Beam skill on player move (experimental)
This is the initial testing approach i took for skill testing.
It does not really seem to be as effective, because having to wait for
the matched by the player line to disappear introduces both a delay and
a tile flow that make it harder to decide if the skill has the desired
visual/gameplay feeling to it.